serveur fini mais jsp si il marche xd

+++ Client-GuitareHero-Ravel/client.h
@@ -15,7 +15,7 @@ public:
         QTcpSocket *getSockClient() const;
 
     private:
-        int scoreJoueur;
+        float scoreJoueur;
         QString pseudo;
         QTcpSocket *sockClient;
 };
